Intermediate response to feedback
Mostly just better variable names and method names.
