+++ tests/python/pants_test/android/tasks/test_aapt_task.py
@@ -18,8 +18,17 @@ class TestAaptTask(TestAndroidBase):
   def task_type(cls):
     return AaptTask
 
-  # This just tests the AaptTask classmethods. Any methods that need a task instance
+  # This just tests the AaptTask static/class methods. Any methods that need a task instance
   # are exercised in tests for AaptTask subclasses.
+
+  def test_is_aapt_target(self):
+    with self.android_binary() as android_binary:
+      self.assertTrue(AaptTask.is_android_binary(android_binary))
+
+  def test_not_aapt_target(self):
+    with self.android_library() as android_library:
+      self.assertFalse(AaptTask.is_android_binary(android_library))
+
   def test_package_path_translation(self):
     self.assertEqual(os.path.join('com', 'pants', 'example', 'tests'),
                      AaptTask.package_path('com.pants.example.tests'))
